Refactoring file system methods (work in progress)
