+++ src/commands/lambda/flare.ts
@@ -25,6 +25,7 @@ import {
   DATADOG_SITE_US1,
   DATADOG_SITES,
 } from '../../constants'
+import {deleteFolder} from '../../helpers/fileSystem'
 import * as helpersRenderer from '../../helpers/renderer'
 import {isValidDatadogSite} from '../../helpers/validation'
 
@@ -348,21 +349,6 @@ export const maskConfig = (config: FunctionConfiguration) => {
   const stringifiedConfig = JSON.stringify(config, replacer)
 
   return JSON.parse(stringifiedConfig) as FunctionConfiguration
-}
-
-/**
- * Delete a folder and all its contents
- * @param folderPath the folder to delete
- * @throws Error if the deletion fails
- */
-export const deleteFolder = (folderPath: string) => {
-  try {
-    fs.rmSync(folderPath, {recursive: true, force: true})
-  } catch (err) {
-    if (err instanceof Error) {
-      throw Error(`Failed to delete files located at ${folderPath}: ${err.message}`)
-    }
-  }
 }
 
 /**
