+++ src/helpers/fileSystem.ts
@@ -1,0 +1,16 @@
+import fs from 'fs'
+
+/**
+ * Delete a folder and all its contents
+ * @param folderPath the folder to delete
+ * @throws Error if the deletion fails
+ */
+export const deleteFolder = (folderPath: string) => {
+  try {
+    fs.rmSync(folderPath, {recursive: true, force: true})
+  } catch (err) {
+    if (err instanceof Error) {
+      throw Error(`Failed to delete files located at ${folderPath}: ${err.message}`)
+    }
+  }
+}
